requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0

# API and HTTP
requests>=2.28.2
httpx>=0.24.0

# AI/ML
google-generativeai>=0.3.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Utilities
pydantic>=1.10.7
asyncio>=3.4.3
json5>=0.9.14